ci: Retry downloads over Tor
We have seen that our CI checks fail a few times within a week when
downloading site content over Tor. The landing page seems to operate
properly, so our assumption is that the Tor network may be experiencing
an issue.

Add a CI action that retries the download after 3 minutes, to minimize
the number of false failures.